+++ blabpy/eaf.py
@@ -8,6 +8,10 @@ Notes:
 - It would be cleaner to wrap the XML-tree-based functions in classes at some point.
 - Many functions are XML-general and could be moved to a separate module or aggregated in a separate class.
 
+Glossary:
+- *Aligned* vs. *referenced* annotations: Aligned annotations are the ones that have their own timestamps, e.g.,
+  participant utterances. Referenced don't have their own timestamps, e.g., xds, lex, vcm, etc.
+- *Daughter* vs. *parent* annotations: Tiers are organized in a hierachy
 
 TODO:
 [ ] Refactor add_* functions to use a single add_element function.
@@ -24,6 +28,7 @@ import pandas as pd
 import requests
 from pympi import Eaf
 
+from blabpy.utils import concatenate_dataframes
 
 # Element names
 EXTERNAL_REF = 'EXTERNAL_REF'
@@ -55,6 +60,9 @@ class EafPlus(Eaf):
     """
     This class is just pympi.Eaf plus a few extra methods.
     """
+
+    class AnnotationExtractionError(Exception):
+        pass
 
     def get_time_intervals(self, tier_id):
         """
@@ -101,22 +109,25 @@ class EafPlus(Eaf):
         return participant_tier_ids
 
     def _get_aligned_annotations(self, tier_id):
+        """
+        Get aligned annotations for a given tier.
+        :param tier_id: tier id, e.g., CHI
+        :return: a dataframe with columns onset, offset, annotation, annotation_id
+        """
+        # Load times and annotations. If there aren't any, substitute placeholder to return a DataFrame with all Nones
+        # in the end.
+        # Times
         time_intervals = self.get_time_intervals(tier_id=tier_id)
-        if len(time_intervals) == 0:
-            return None
-
-        onsets, offsets = zip(*time_intervals)
+        onsets, offsets = zip(*time_intervals) if time_intervals else ([None], [None])
+        # Annotations
         aligned_annotations, _, _, _ = self.tiers[tier_id]
-        if aligned_annotations:
-            ids, annotations = zip(*[(id_, annotation)
-                                     for id_, (_, _, annotation, _)
-                                     in aligned_annotations.items()])
-            return pd.DataFrame.from_dict(dict(
-                onset=onsets, offset=offsets,
-                annotation=annotations,
-                annotation_id=ids))
-        else:
-            return pd.DataFrame(columns=['onset', 'offset', 'annotation', 'annotation_id'])
+        aligned_annotations = aligned_annotations or {None: (None, None, None, None)}
+        ids, annotations = zip(*[(id_, annotation) for id_, (_, _, annotation, _) in aligned_annotations.items()])
+
+        return pd.DataFrame.from_dict(dict(
+            onset=onsets, offset=offsets,
+            annotation=annotations,
+            annotation_id=ids))
 
     def _get_reference_annotations(self, tier_id):
         _, reference_annotations, _, _ = self.tiers[tier_id]
@@ -142,30 +153,155 @@ class EafPlus(Eaf):
         if annotations_df is None:
             return None
 
-        # as we add daughter tiers, deepest_annotation_id column will change to their ids
-        annotations_df = annotations_df.rename(columns={'annotation_id': 'deepest_annotation_id'})
+        # Strip white space from annotations
+        annotations_df['annotation'] = annotations_df['annotation'].str.strip()
+
+        # I want to keep the annotation IDs to have a unique identifier. We'll be merging annotations_df with
+        # annotations from daughter tiers which will have their own 'annotation_id' column so we'll rename this one.
+        annotations_df = annotations_df.rename(columns={'annotation_id': 'participant_annotation_id'})
+
+        # Save the number of annotations to check that we don't duplicate any when we later merge them with daughter
+        # tier annotations.
         n_annotations = annotations_df.shape[0]
 
-        if tier_id == 'CHI':
-            daughter_tier_types = ('vcm', 'lex', 'mwu')
-        else:
-            daughter_tier_types = ('xds', )
+        # The annotations are in daughter tiers of the participant tier. Their IDs are in the format "xds@CHI".
+        daughter_tier_ids = [tier_id_ for tier_id_ in self.tiers if tier_id_.endswith(f'@{tier_id}')]
+        if len(daughter_tier_ids) == 0:
+            return annotations_df
 
-        for daughter_tier_type in daughter_tier_types:
-            daughter_tier_id = f'{daughter_tier_type}@{tier_id}'
-            daughter_annotations_df = self._get_reference_annotations(tier_id=daughter_tier_id)
-            daughter_annotations_df = daughter_annotations_df.rename(columns={'annotation': daughter_tier_type})
-            # Merge with previously extracted annotations
-            annotations_df = (annotations_df.merge(daughter_annotations_df,
-                                 how='left',
-                                 left_on='deepest_annotation_id',
-                                 right_on='parent_annotation_id')
-             .drop(columns=['deepest_annotation_id', 'parent_annotation_id'])
-             .rename(columns={'annotation_id': 'deepest_annotation_id'})
-             )
+        # Gather all annotation from daughter tiers
+        daughter_annotations = concatenate_dataframes(
+            dataframes=[self._get_reference_annotations(tier_id=daughter_tier_id)
+                        for daughter_tier_id in daughter_tier_ids],
+            keys=daughter_tier_ids,
+            key_column_name='daughter_tier_id')
 
-        assert annotations_df.shape[0] == n_annotations
-        return annotations_df.drop(columns=['deepest_annotation_id'])
+        # If there aren't any daughter annotations, we are done.
+        if daughter_annotations.shape[0] == 0:
+            return annotations_df
+
+        # Empty annotations ('') do not represent anything meaningful: they are just there to be filled by an annotator
+        # later.
+        daughter_annotations = daughter_annotations[daughter_annotations['annotation'] != '']
+
+        # Strip white space from annotations
+        daughter_annotations['annotation'] = daughter_annotations['annotation'].str.strip()
+
+        # Now, we are going to merge the participant annotations (annotations_df) with the daughter annotations
+        # (daughter_annotations) iteratively. Each time, we are going to add all daughter tiers one level deeper
+        # down the hierarchy. There can't be more levels that there are daughter tiers, so that's how many times we will
+        # try adding more annotations by merging. Each merge will add as many columns as there are daughter tiers on the
+        # current level - one column per daughter tier.
+        # For example, if we have the following hierarchy:
+        # FA1 -> x -> y -> z
+        #     \
+        #      -> a -> b
+        # then the three meaningful merges will add columns (x, a), (y, b), (z).
+        annotations_df['deepest_annotation_id'] = annotations_df['participant_annotation_id']
+        daughter_annotations = daughter_annotations.rename(columns={'annotation': 'daughter_annotation'})
+        for level in range(len(daughter_tier_ids)):
+            annotations_df = (
+                annotations_df
+                .merge(
+                    daughter_annotations,
+                    how='left',
+                    left_on='deepest_annotation_id',
+                    right_on='parent_annotation_id',)
+                .drop(columns=['deepest_annotation_id', 'parent_annotation_id'])
+                .rename(columns={f'annotation_id': 'daughter_annotation_id'})
+                .assign(**{'deepest_annotation_id': lambda df: df['daughter_annotation_id']})
+                )
+
+            # Due to parallel daughter tier branches, we might need fewer iterations than there are daughter tiers. In
+            # that case, we'll have NaNs in the deepest_annotation_id column. We'll remove the empty columns we've just
+            # added and stop the loop.
+            just_added_columns = 'daughter_tier_id', 'daughter_annotation', 'daughter_annotation_id'
+            if annotations_df['deepest_annotation_id'].isna().all():
+                annotations_df = annotations_df.drop(columns=list(just_added_columns))
+                last_level = level - 1
+                break
+
+            else:
+                # We are adding `just_added_columns` every time, so we'll add a suffix to distinguish them.
+                suffix = f'_{level}'
+                annotations_df = annotations_df.rename(columns={col_name: f'{col_name}{suffix}'
+                                                                for col_name in just_added_columns})
+                last_level = level
+
+        # We only need annotations IDs from one level to keep track of parallel annotations one level deeper. There is
+        # no "deeper" for the last level, so we can drop its IDs.
+        annotations_df.drop(columns=[f'daughter_annotation_id_{last_level}', 'deepest_annotation_id'], inplace=True)
+
+        # ???
+        # We have to go in reverse order because ??? ~otherwise we'll lose parent_annotation_id in case of parallel tier branches~
+        for level in reversed(range(last_level + 1)):
+            # We need to keep track
+            if level == 0:
+                parent_annotation_id_column = 'participant_annotation_id'
+            else:
+                parent_annotation_id_column = f'daughter_annotation_id_{level - 1}'
+            annotations_df = (
+                annotations_df
+                # We need to pivot daughter tier IDs and annotation into columns and their values respectively.
+                # Everything else should at most collapse in case of parallel daughter annotation of the same parent
+                # annotation and should otherwise stay the same. Importantly, everything else includes the parent
+                # annotation ID so that parallel daughter annotations end up in the same row. To keep everything else
+                # intact, we'll set it those columns as index first.
+                .set_index(
+                        [c for c in annotations_df.columns.values if
+                         not (c.startswith('daughter_') and c.endswith(f'_{level}'))])
+                # Pivot 'daughter_tier_id', 'daughter_annotation' using `parent_annotation_id` as row ID.
+                .set_index([f'daughter_tier_id_{level}'], append=True)
+                .unstack(level=-1)
+                # Unstack saves the names of the pivoted columns in the columns index. We don't need that reminder.
+                .droplevel(level=0, axis=1)
+                .rename_axis(None, axis=1)
+                .reset_index(drop=False)
+                # Drop the parent annotation IDs - they are the deepest level IDs now, and we don't need them anymore.
+                .drop(columns=(list() if level == 0 else [parent_annotation_id_column]))
+                # Some annotations might not have had any daughter annotations at the level we've just unstacked. In
+                # that case, we'll have an <NA> column. It will be completely empty, because if daughter annotation
+                # IDs are missing, then the corresponding daughter annotations are missing too. We'll drop the column.
+                .drop(columns=pd.NA, errors='ignore')
+            )
+
+        # The comments might make it seem that we should be done by this point, but we are not. We have all the columns
+        # and all the data but there might still be multiple rows per annotation. Here is an example:
+        # 'baby', 'a1' - speaker-level annotation.
+        # daughter annotations:
+        # 'tier_id', 'ann_id', 'parent_ann_id', 'annotation'
+        # 'x21@FA1', 'a21',    'a1',            'A'
+        # 'x22@FA1', 'a22',    'a1',            'B'
+        # 'x31@FA1', 'a31',    'a21'            'C'
+        # Merging result:
+        # 'ann', 'speaker_ann_id', 'tier_id_0', 'ann_0', 'ann_id_0', 'tier_id_1', 'ann_1', 'ann_id_1'
+        # 'baby', 'a1',            'x21@FA1',   'A',     'a21',      'x31@FA1',   'C',     'a31'
+        # 'baby', 'a1',            'x22@FA1',   'B',     'a22',      <NA>,        <NA>,    <NA>
+        # Unstacking result:
+        # 'ann', 'speaker_ann_id', 'x31@FA1', 'x21@FA1', 'x22@FA1'
+        # 'baby', 'a1',            C,         'A',       '<NA>'
+        # 'baby', 'a1',            <NA>,      <NA>,       'B'
+        # We need to collapse the rows with the same annotation and speaker_ann_id. We'll do that by grouping by
+        # annotation and speaker_ann_id and concatenating the values in the other columns.
+        non_daughter_annotation_columns = [c for c in annotations_df.columns.values if '@' not in c]
+        annotations_df = (
+            annotations_df
+            .set_index(non_daughter_annotation_columns)
+            .groupby(non_daughter_annotation_columns)
+            .transform(lambda x: sorted(x, key=lambda k: pd.isna(k)))
+            .groupby(non_daughter_annotation_columns)
+            .last()
+            .reset_index(drop=False)
+        )
+
+        # Remove the suffixes from the column names: xds@FA1 -> xds
+        annotations_df = annotations_df.rename(columns={col_name: col_name.split('@')[0]
+                                                        for col_name in annotations_df.columns.values})
+
+        if annotations_df.shape[0] != n_annotations:
+            raise self.AnnotationExtractionError('The number of annotations has changed during the extraction process.')
+
+        return annotations_df
 
     def get_full_annotations(self):
         """