+++ blabpy/vihi/tests/test_paths.py
@@ -1,7 +1,7 @@
 from blabpy.vihi.paths import get_vihi_path, get_subject_files_path, get_lena_path, get_lena_annotations_path, \
     compose_full_recording_id, parse_full_recording_id, get_lena_population_path, get_lena_subject_path, \
     get_lena_recording_path, get_raw_data_dir, get_its_dir, get_its_path, get_rttm_path, get_eaf_path, \
-    get_lena_annotations_in_progress_path
+    get_lena_annotations_in_progress_path, find_all_lena_recording_ids
 
 
 def _test_get_path_function(function, name, *args, **kwargs):
@@ -35,3 +35,9 @@ def test_parse_full_recording_id():
     assert parse_full_recording_id('TD_123_456') == {'population': 'TD',
                                                      'subject_id': '123',
                                                      'recording_id': '456'}
+
+
+def test_find_all_lena_recording_ids():
+    all_lena_recording_ids = find_all_lena_recording_ids()
+    assert type(all_lena_recording_ids) == list
+    assert all(type(recording_id) == str for recording_id in all_lena_recording_ids)